requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.29
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
python-slugify==8.0.4